+++ player/mixer.c
@@ -113,7 +113,7 @@
 	position = chan->position_frac; \
 	const int##bits##_t *p = (int##bits##_t *)(chan->current_sample_data + (chan->position * (bits / 8))); \
 	if (chan->flags & CHN_STEREO) p += chan->position; \
-	int *pvol = pbuffer;\
+	int32_t *pvol = pbuffer;\
 	do {
 
 
@@ -308,11 +308,11 @@
 //////////////////////////////////////////////////////////
 // Interfaces
 
-typedef void(* mix_interface_t)(song_voice_t *, int *, int *);
+typedef void(* mix_interface_t)(song_voice_t *, int32_t *, int32_t *);
 
 
 #define BEGIN_MIX_INTERFACE(func) \
-	static void func(song_voice_t *channel, int *pbuffer, int *pbufmax) \
+	static void func(song_voice_t *channel, int32_t *pbuffer, int32_t *pbufmax) \
 	{ \
 		int_fast32_t position;
 
@@ -556,10 +556,10 @@ static const mix_interface_t fastmix_functions[2 * 2 * 16] = {
 	BUILD_MIX_FUNCTION_TABLE_FAST(FirFilter)
 };
 
-static int get_sample_count(song_voice_t *chan, int samples)
+static int get_sample_count(song_voice_t *chan, int32_t samples)
 {
-	int loop_start = (chan->flags & CHN_LOOP) ? chan->loop_start : 0;
-	int increment = chan->increment;
+	int32_t loop_start = (chan->flags & CHN_LOOP) ? chan->loop_start : 0;
+	int32_t increment = chan->increment;
 
 	if (samples <= 0 || !increment || !chan->length)
 		return 0;
@@ -572,7 +572,7 @@ static int get_sample_count(song_voice_t *chan, int samples)
 			chan->position = loop_start + (delta >> 16);
 			chan->position_frac = delta & 0xFFFF;
 
-			if ((int) chan->position < loop_start ||
+			if ((int32_t) chan->position < loop_start ||
 				chan->position >= (loop_start + chan->length) / 2) {
 				chan->position = loop_start;
 				chan->position_frac = 0;
@@ -592,7 +592,7 @@ static int get_sample_count(song_voice_t *chan, int samples)
 		}
 		else {
 			// We probably didn't hit the loop end yet (first loop), so we do nothing
-			if ((int) chan->position < 0)
+			if ((int32_t) chan->position < 0)
 				chan->position = 0;
 		}
 	}
@@ -638,7 +638,7 @@ static int get_sample_count(song_voice_t *chan, int samples)
 		}
 	}
 
-	int position = chan->position;
+	int32_t position = chan->position;
 
 	// too big increment, and/or too small loop length
 	if (position < loop_start) {
@@ -649,12 +649,12 @@ static int get_sample_count(song_voice_t *chan, int samples)
 	if (position < 0 || position >= (int) chan->length)
 		return 0;
 
-	int position_frac = (unsigned short) chan->position_frac,
+	int32_t position_frac = (uint16_t) chan->position_frac,
 		 sample_count = samples;
 
 	if (increment < 0) {
-		int inv = -increment;
-		int maxsamples = 16384 / ((inv >> 16) + 1);
+		int32_t inv = -increment;
+		int32_t maxsamples = 16384 / ((inv >> 16) + 1);
 
 		if (maxsamples < 2)
 			maxsamples = 2;
@@ -662,19 +662,19 @@ static int get_sample_count(song_voice_t *chan, int samples)
 		if (samples > maxsamples)
 			samples = maxsamples;
 
-		int delta_hi = (inv >> 16) * (samples - 1);
-		int delta_lo = (inv & 0xffff) * (samples - 1);
-		int pos_dest = position - delta_hi + ((position_frac - delta_lo) >> 16);
+		int32_t delta_hi = (inv >> 16) * (samples - 1);
+		int32_t delta_lo = (inv & 0xffff) * (samples - 1);
+		int32_t pos_dest = position - delta_hi + ((position_frac - delta_lo) >> 16);
 
 		if (pos_dest < loop_start) {
 			sample_count =
-				(unsigned int) (((((long long) position -
+				(uint32_t) (((((int64_t) position -
 					loop_start) << 16) + position_frac -
 					  1) / inv) + 1;
 		}
 	}
 	else {
-		int maxsamples = 16384 / ((increment >> 16) + 1);
+		int32_t maxsamples = 16384 / ((increment >> 16) + 1);
 
 		if (maxsamples < 2)
 			maxsamples = 2;
@@ -682,13 +682,13 @@ static int get_sample_count(song_voice_t *chan, int samples)
 		if (samples > maxsamples)
 			samples = maxsamples;
 
-		int delta_hi = (increment >> 16) * (samples - 1);
-		int delta_lo = (increment & 0xffff) * (samples - 1);
-		int pos_dest = position + delta_hi + ((position_frac + delta_lo) >> 16);
+		int32_t delta_hi = (increment >> 16) * (samples - 1);
+		int32_t delta_lo = (increment & 0xffff) * (samples - 1);
+		int32_t pos_dest = position + delta_hi + ((position_frac + delta_lo) >> 16);
 
-		if (pos_dest >= (int) chan->length) {
-			sample_count = (unsigned int)
-				(((((long long) chan->length - position) << 16) - position_frac - 1) / increment) + 1;
+		if (pos_dest >= (int32_t) chan->length) {
+			sample_count = (uint32_t)
+				(((((int64_t) chan->length - position) << 16) - position_frac - 1) / increment) + 1;
 		}
 	}
 
@@ -701,9 +701,9 @@ static int get_sample_count(song_voice_t *chan, int samples)
 }
 
 
-unsigned int csf_create_stereo_mix(song_t *csf, int count)
+uint32_t csf_create_stereo_mix(song_t *csf, uint32_t count)
 {
-	int* ofsl, *ofsr;
+	int32_t* ofsl, *ofsr;
 	unsigned int nchused, nchmixed;
 
 	if (!count)
@@ -713,17 +713,17 @@ unsigned int csf_create_stereo_mix(song_t *csf, int count)
 
 	// yuck
 	if (csf->multi_write)
-		for (unsigned int nchan = 0; nchan < MAX_CHANNELS; nchan++)
+		for (uint32_t nchan = 0; nchan < MAX_CHANNELS; nchan++)
 			memset(csf->multi_write[nchan].buffer, 0, sizeof(csf->multi_write[nchan].buffer));
 
-	for (unsigned int nchan = 0; nchan < csf->num_voices; nchan++) {
+	for (uint32_t nchan = 0; nchan < csf->num_voices; nchan++) {
 		const mix_interface_t *mix_func_table;
 		song_voice_t *const channel = &csf->voices[csf->voice_mix[nchan]];
-		unsigned int flags;
-		unsigned int nrampsamples;
-		int smpcount;
-		int nsamples;
-		int *pbuffer;
+		uint32_t flags;
+		uint32_t nrampsamples;
+		int32_t smpcount;
+		int32_t nsamples;
+		int32_t *pbuffer;
 
 		if (!channel->current_sample_data)
 			continue;
@@ -776,13 +776,13 @@ unsigned int csf_create_stereo_mix(song_t *csf, int count)
 
 		nchused++;
 		////////////////////////////////////////////////////
-		unsigned int naddmix = 0;
+		uint32_t naddmix = 0;
 
 		do {
 			nrampsamples = nsamples;
 
 			if (channel->ramp_length > 0) {
-				if ((int) nrampsamples > channel->ramp_length)
+				if ((int32_t) nrampsamples > channel->ramp_length)
 					nrampsamples = channel->ramp_length;
 			}
 
@@ -815,7 +815,7 @@ unsigned int csf_create_stereo_mix(song_t *csf, int count)
 
 			if ((nchmixed >= max_voices && !(csf->mix_flags & SNDMIX_DIRECTTODISK))
 				|| (!channel->ramp_length && !(channel->left_volume | channel->right_volume))) {
-				int delta = (channel->increment * (int) smpcount) + (int) channel->position_frac;
+				int32_t delta = (channel->increment * (int32_t) smpcount) + (int32_t) channel->position_frac;
 				channel->position_frac = delta & 0xFFFF;
 				channel->position += (delta >> 16);
 				channel->rofs = channel->lofs = 0;
@@ -830,7 +830,7 @@ unsigned int csf_create_stereo_mix(song_t *csf, int count)
 					mix_func = channel->ramp_length
 						? mix_func_table[flags | MIXNDX_RAMP]
 						: mix_func_table[flags];
-					int *pbufmax = pbuffer + (smpcount * 2);
+					int32_t *pbufmax = pbuffer + (smpcount * 2);
 					channel->rofs = -*(pbufmax - 2);
 					channel->lofs = -*(pbufmax - 1);
